requirements.txt
# Core scientific computing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
statsmodels>=0.13.0
hmmlearn>=0.2.7

# Interpretability
shap>=0.41.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Data formats
pyarrow>=8.0.0
tables>=3.7.0

# Hawkes processes 
# tick==0.6.0.0  # Can be difficult to install, commented out by default

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
python-dotenv>=0.19.0

# API/Data collection
requests>=2.26.0
websocket-client>=1.2.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality (optional)
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0